+++ docs/changelog.rst
@@ -1,4 +1,25 @@
-.. currentmodule:: twitchio
+:orphan:
+
+2.2.0
+=====
+- ext.sounds
+    - Added sounds extension. Check the sounds documentation for more information.
+
+- TwitchIO
+    - Loosen aiohttp requirements to allow 3.8.1
+    - :class:`Stream` was missing from ``__all__``. It is now available in the twitchio namespace.
+    - Added ``.status``, ``.reason`` and ``.extra`` to :class:`HTTPException`
+    - fix Message._timestamp value when tag is not provided by twitch
+  
+- ext.pubsub
+    - Add channel subscription pubsub model.
+
+2.1.5
+=====
+- TwitchIO
+    - Add ``user_id`` property to Client
+    - Change id_cache to only cache if a value is not ``None``
+    - Add :func:`Client.wait_for_ready`
 
 Master
 =======